refactor(plugins): keep prepared index contracts acyclic
